ユーザー情報登録画面まで実装1114

+++ reservation/src/main/java/com/example/demo/repository/PlanRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.entity.PlanEntity;
 
 @Repository
-public interface PlanRepository extends JpaRepository<PlanEntity, Integer> {
+public interface PlanRepository extends JpaRepository<PlanEntity, Long> {
 }
